OCG System GR/requirements.txt
fastapi
uvicorn
sqlalchemy
argon2-cffi
pydantic
streamlit
requests
pandas
plotly
Procfile